requirements.txt
pyyaml==6.0
boto3==1.24.64
-e git+https://github.com/yigitbasalma/bunch.git#egg=bunch